Remove more circular dependencies with ActuatorEffectiveness
