+++ src/lib/CMakeLists.txt
@@ -31,6 +31,7 @@
 #
 ############################################################################
 
+add_subdirectory(actuator_effectiveness EXCLUDE_FROM_ALL)
 add_subdirectory(adsb EXCLUDE_FROM_ALL)
 add_subdirectory(airspeed EXCLUDE_FROM_ALL)
 add_subdirectory(atmosphere EXCLUDE_FROM_ALL)
@@ -43,6 +44,7 @@ add_subdirectory(cdrstream EXCLUDE_FROM_ALL)
 add_subdirectory(circuit_breaker EXCLUDE_FROM_ALL)
 add_subdirectory(collision_prevention EXCLUDE_FROM_ALL)
 add_subdirectory(component_information EXCLUDE_FROM_ALL)
+add_subdirectory(control_allocation EXCLUDE_FROM_ALL)
 add_subdirectory(controllib EXCLUDE_FROM_ALL)
 add_subdirectory(conversion EXCLUDE_FROM_ALL)
 add_subdirectory(crc EXCLUDE_FROM_ALL)
