+++ src/lib/actuator_effectiveness/ActuatorEffectivenessFixedWing.hpp
