+++ src/lib/actuator_effectiveness/ActuatorEffectivenessHelicopter.cpp
