+++ src/lib/actuator_effectiveness/ActuatorEffectivenessHelicopter.hpp
