+++ src/lib/actuator_effectiveness/ActuatorEffectivenessHelicopterCoaxial.cpp
