+++ src/lib/actuator_effectiveness/ActuatorEffectivenessHelicopterCoaxial.hpp
