+++ src/lib/actuator_effectiveness/ActuatorEffectivenessRotors.cpp
