+++ src/lib/actuator_effectiveness/ActuatorEffectivenessRoverAckermann.hpp
