+++ src/modules/control_allocator/CMakeLists.txt
@@ -32,8 +32,6 @@
 ############################################################################
 
 include_directories(${CMAKE_CURRENT_SOURCE_DIR})
-add_subdirectory(ActuatorEffectiveness)
-add_subdirectory(ControlAllocation)
 
 px4_add_module(
 	MODULE modules__control_allocator
